STYLE: fix warning about int/float
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=4393
git-svn-id: http://svn.slicer.org/Slicer4/trunk@4393 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Base/GUI/vtkSlicerViewControlGUI.cxx
@@ -1454,8 +1454,8 @@ void vtkSlicerViewControlGUI::DeviceCoordinatesToXYZ(vtkSlicerSliceGUI *sgui, in
 
   xyz[0] = x - origin[0];
   xyz[1] = y - origin[1];
-  xyz[2] = floor(ty*snode->GetLayoutGridRows())*snode->GetLayoutGridColumns()
-    + floor(tx*snode->GetLayoutGridColumns());
+  xyz[2] = static_cast<int> ( floor(ty*snode->GetLayoutGridRows())*snode->GetLayoutGridColumns()
+      + floor(tx*snode->GetLayoutGridColumns()) );
 }
